CEDS-942 Move Submission related model classes to submissions package
