+++ app/uk/gov/hmrc/exports/models/declaration/submissions/RequestType.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.exports.models.declaration
+package uk.gov.hmrc.exports.models.declaration.submissions
 
 import play.api.libs.json._
