+++ test/integration/uk/gov/hmrc/exports/connector/CustomsDeclarationsConnectorSpec.scala
@@ -24,7 +24,7 @@ import play.api.Application
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.test.Helpers._
 import uk.gov.hmrc.exports.connectors.CustomsDeclarationsConnector
-import uk.gov.hmrc.exports.models.declaration.CustomsDeclarationsResponse
+import uk.gov.hmrc.exports.models.CustomsDeclarationsResponse
 import uk.gov.hmrc.http.HeaderCarrier
 import util.CustomsDeclarationsAPIConfig
 import util.ExternalServicesConfig.{Host, Port}
